% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vamf.R
\name{colNorms}
\alias{colNorms}
\title{L2 norms of columns of a matrix.}
\usage{
colNorms(x)
}
\arguments{
\item{x}{a matrix}
}
\value{
a vector containing the L2 norms of the columns of x.
}
\description{
L2 norms of columns of a matrix.
}
