% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vamf.R
\name{effective_dimension}
\alias{effective_dimension}
\title{Effective dimension.
Finds the effective dimension of a latent factor matrix by first computing L2 norms of each column.
Columns with L2 norm greater than the maximum norm times 'thresh' are nonzero.
Columns with L2 norm less than the cutoff (above) are 'zero'.}
\usage{
effective_dimension(x, thresh = 0.05)
}
\arguments{
\item{x}{a matrix whose columns represent coordinates in a latent space}

\item{thresh}{a threshold close to zero}
}
\value{
a count of the number of nonzero columns
}
\description{
Effective dimension.
Finds the effective dimension of a latent factor matrix by first computing L2 norms of each column.
Columns with L2 norm greater than the maximum norm times 'thresh' are nonzero.
Columns with L2 norm less than the cutoff (above) are 'zero'.
}
