% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vamf.R
\name{extract_elbo}
\alias{extract_elbo}
\title{Extract evidence lower bound (ELBO).}
\usage{
extract_elbo(logtxt)
}
\arguments{
\item{logtxt}{A list of strings. Each element is a single line of the output log text of a Stan run.}
}
\value{
The scalar value of the ELBO.
}
\description{
Takes the log text from running variational Bayes in Stan and extracts out the value of the evidence lower bound (ELBO) at the final iteration. Higher ELBO is better.
}
