% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vamf.R
\name{extract_factors}
\alias{extract_factors}
\title{Extract latent factors from a Stan model fit.
Extracts out the posterior means of the latent factors and all other parameter values from a Stan model fit.}
\usage{
extract_factors(stan_vb_fit, ss)
}
\arguments{
\item{stan_vb_fit}{Stan variational Bayes object obtained from the output of \code{\link{vb_wrap}}.}

\item{ss}{List of sufficient statistics from \code{\link{init_ss}}.}
}
\value{
List of posterior means of all model parameters.
}
\description{
Extract latent factors from a Stan model fit.
Extracts out the posterior means of the latent factors and all other parameter values from a Stan model fit.
}
