% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vamf.R
\name{norm}
\alias{norm}
\title{L2 norm of an object.
L2 norm of an object.}
\usage{
norm(v)
}
\arguments{
\item{v}{An vector or array-like object}
}
\value{
The L2 (Euclidean) norm of v
}
\description{
L2 norm of an object.
L2 norm of an object.
}
