% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vamf.R
\name{ortho_vamf}
\alias{ortho_vamf}
\title{Orthogonalize and normalize latent factors.}
\usage{
ortho_vamf(vars)
}
\arguments{
\item{vars}{List of posterior means of model parameters from \code{\link{extract_factors}}.}
}
\value{
The same list but with additional components 'factors' and 'loadings' whose interpretation is analogous to Principal Components Analysis.
}
\description{
Rotate and transform the latent factors to an equivalent representation that facilitates interpretability by using an orthonormal basis in the loading matrix.
}
