% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vamf.R
\name{rm_zero_rowcol}
\alias{rm_zero_rowcol}
\title{Remove zero rows and columns.
Removes all rows and columns that have only zeros from a sparse Matrix}
\usage{
rm_zero_rowcol(Y)
}
\arguments{
\item{Y}{A sparse Matrix}
}
\value{
The same Matrix without zero rows and columns
}
\description{
Remove zero rows and columns.
Removes all rows and columns that have only zeros from a sparse Matrix
}
