% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vamf.R
\name{vamf}
\alias{vamf}
\title{Varying-Censoring Aware Matrix Factorization.}
\usage{
vamf(Y, L, nrestarts = 4, log2trans = TRUE, pseudocount = 0,
  output_samples = 100, save_restarts = FALSE, svmult = 1)
}
\arguments{
\item{Y}{Sparse Matrix of gene expression measurements, with genetic features (genes) in the rows and samples (typically, individual cells) in the columns.}

\item{L}{Upper bound on the dimensionality of the latent space to be learned. Automatic relevance determination is used to shrink away unnecessary dimensions.}

\item{nrestarts}{Number of independent random initializations of the algorithm. Can be parallelized by setting e.g. \code{options(mc.cores=4)}.}

\item{log2trans}{Should the data be log transformed prior to analysis? Set to FALSE if the data have already been log transformed.}

\item{pseudocount}{Optional small offset to be added to data before log transformation.}

\item{output_samples}{Number of samples from approximate posterior used to estimate the posterior means of all parameters.}

\item{save_restarts}{If multiple initializations are used, set this to TRUE if you want to return the list of all results. Set to FALSE to choose only the best result based on the highest evidence lower bound (ELBO).}

\item{svmult}{Scalar or vector of multipliers to increase or decrease the sigma_v scale hyperparameter.}
}
\value{
List of posterior means of all model parameters. The 'factors' and 'loadings' are analogous to PCA. Cell positions in latent space can be plotted by using the 'factors' matrix.
}
\description{
VAMF is a probabilistic dimension reduction method intended for single cell RNA-Seq datasets.
}
