% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vamf.R
\name{vamf_stan}
\alias{vamf_stan}
\title{Varying-Censoring Aware Matrix Factorization via Stan.
Runs VAMF with many random initializations in parallel. Parallelism will only work if the number of cores is set as a global option. For example \code{options(mc.cores=4)}. Discards any random restarts that resulted in Stan errors, usually due to numeric convergence failures.}
\usage{
vamf_stan(ss, svmult = rep.int(1, 4), output_samples = 100)
}
\arguments{
\item{ss}{A list of sufficient statistics from \code{\link{init_ss}}.}

\item{svmult}{Vector of multipliers to increase or decrease the sigma_v scale hyperparameter. Length of svmult determines number of random restarts to run.}

\item{output_samples}{How many samples from the approximate posterior to estimate the posterior mean of each parameter.}
}
\value{
List of results from each random initialization excluding numerical convergence failures. Each element of the list is an object with structure described in \code{\link{vb_wrap}}.
}
\description{
Varying-Censoring Aware Matrix Factorization via Stan.
Runs VAMF with many random initializations in parallel. Parallelism will only work if the number of cores is set as a global option. For example \code{options(mc.cores=4)}. Discards any random restarts that resulted in Stan errors, usually due to numeric convergence failures.
}
