% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vamf.R
\name{vb_wrap}
\alias{vb_wrap}
\title{Variational Bayes wrapper function.
Convenience function for parallelizing calls to VB.}
\usage{
vb_wrap(svmult, stmod, ss, resnames, output_samples)
}
\arguments{
\item{svmult}{Scalar multiplier to increase or decrease the sigma_v scale hyperparameter.}

\item{stmod}{An object of class \code{\linkS4class{stanmodel}}.}

\item{ss}{A list of sufficient statistics from \code{\link{init_ss}}.}

\item{resnames}{A string vector of parameter names to be included in output.}

\item{output_samples}{How many samples from the approximate posterior to estimate the posterior mean of each parameter.}
}
\value{
A list with components stan_vb_fit (the result from Stan) and logtxt (the log of Stan output).
}
\description{
Variational Bayes wrapper function.
Convenience function for parallelizing calls to VB.
}
